backend/requirements.txt
Django>=2.2,<2.3
djangorestframework>=3.11,<3.12

psycopg2-binary>=2.8,<2.9

Pillow>=7.1.0,<8

django-cors-headers
gunicorn>=20.0.4,<21